+++ taze.config.mjs
@@ -1,0 +1,9 @@
+import { defineConfig } from 'taze'
+
+export default defineConfig({
+  exclude: [
+    'typescript',
+    'markdown-it',
+    'linkify-it',
+  ],
+})